refactor: rename Icon component
